requirements.txt
# Backend requirements
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
opencv-python==4.7.0.72
numpy==1.24.2
pillow==9.4.0
requests==2.28.2
werkzeug==2.2.3

# AI Model APIs
google-generativeai==0.3.0
openai==1.3.0
anthropic==0.5.0

# ML Libraries (for custom model)
torch==2.0.0
torchvision==0.15.1